+++ services/dumper.js
@@ -3,6 +3,7 @@ const fs = require('fs');
 const _ = require('lodash');
 const mkdirpSync = require('mkdirp');
 const chalk = require('chalk');
+const { plural, singular } = require('pluralize');
 const stringUtils = require('../utils/strings');
 const logger = require('./logger');
 
@@ -74,13 +75,6 @@ function Dumper(config) {
     const template = _.template(fs.readFileSync(templatePath, 'utf-8'));
 
     writeFile(`${path}/.gitignore`, template({}));
-  }
-
-  function writeDotGitKeep(pathDest) {
-    const templatePath = `${__dirname}/../templates/app/gitkeep`;
-    const template = _.template(fs.readFileSync(templatePath, 'utf-8'));
-
-    writeFile(`${pathDest}/.gitkeep`, template({}));
   }
 
   function getDatabaseUrl() {
@@ -171,7 +165,25 @@ function Dumper(config) {
       dialect: config.dbDialect,
     });
 
-    writeFile(`${path}/models/${table}.js`, text);
+    const modelNameDasherized = _.kebabCase(table);
+    writeFile(`${path}/models/${modelNameDasherized}.js`, text);
+  }
+
+  function writeRoute(modelName) {
+    const templatePath = `${__dirname}/../templates/app/routes/route.txt`;
+    const template = _.template(fs.readFileSync(templatePath, 'utf-8'));
+
+    const modelNameDasherized = _.kebabCase(modelName);
+    const readableModelName = _.startCase(modelName);
+    const text = template({
+      modelName,
+      modelNameDasherized,
+      modelNameReadablePlural: plural(readableModelName),
+      modelNameReadableSingular: singular(readableModelName),
+      dbDialect: config.dbDialect,
+    });
+
+    writeFile(`${path}/routes/${modelNameDasherized}.js`, text);
   }
 
   function writeForestCollection(table) {
@@ -179,7 +191,8 @@ function Dumper(config) {
     const template = _.template(fs.readFileSync(templatePath, 'utf-8'));
     const text = template({ ...config, table });
 
-    writeFile(`${path}/forest/${table}.js`, text);
+    const modelNameDasherized = _.kebabCase(table);
+    writeFile(`${path}/forest/${modelNameDasherized}.js`, text);
   }
 
   function writeAppJs() {
@@ -276,7 +289,9 @@ function Dumper(config) {
     });
 
     copyTemplate('public/favicon.png', `${path}/public/favicon.png`);
-    writeDotGitKeep(routesPath);
+    modelNames.forEach((modelName) => {
+      writeRoute(modelName);
+    });
     copyTemplate('views/index.html', `${path}/views/index.html`);
 
     writeDotDockerIgnore();